sdm845-common: Add shim for missing symbol in lib-imsvideocodec.so Commit 992496bd2e ("Allow creating child surfaces from BlastBufferQueue") added 3rd parameter to Surface::Surface constructor that causes an unresolved symbol in existing blobs.
Provide a shim library with implementation of older Surface
constructor for lib-imsvidcodec.so that references the mentioned
symbol. Class constructor is a function with thiscall calling
convention so to preserve ABI we have to provide a C-function with
pointer as first parameter (that is hidden in C++) and return that
pointer from the function.

Resolves:
E ims_media_jni: Error loading library lib-imsvt.so: dlopen failed: cannot locate symbol "_ZN7android7SurfaceC1ERKNS_2spINS_22IGraphicBufferProducerEEEb" referenced by "/system/system_ext/lib64/lib-imsvideocodec.so"

Change-Id: I91abeabb07cac8553ee548e507a949e929590995
Signed-off-by: krtonia <krtonia@gmail.com>
Signed-off-by: Ivan Vecera <ivan@cera.cz>